Refactor: move data to data directory
Use the "ducks" pattern (https://github.com/erikras/ducks-modular-redux)
and registering reducers to keep data ducks standalone.

Fix the authentication so that the router works neatly with the auth.
Unauth'd just renders the login page, so there's no redirecting on
failed login, which means the right thing is rendered when login is
successful.